requirements.txt
pymysql
PyJWT
flask
gunicorn
flask-cors
